requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==2.1.1
flake8==6.0.0
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
mccabe==0.7.0
multidict==6.0.3
mypy==0.991
mypy-extensions==0.4.3
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
tomli==2.0.1
types-requests==2.28.11.5
types-urllib3==1.26.25.4
typing-extensions==4.4.0
urllib3==1.26.13
yarl==1.8.2
